coffee_agntcy_supervisor/pyproject.toml
[tool.poetry]
name = "acorda"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "supervisor" }
]

[tool.poetry.dependencies]
python = ">=3.12.4,<4.0.0"
requests = ">=2.32.3,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
langchain-core = "*"
langgraph = "*"
pydantic = "*"
langchain = "^0.3.25"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pydantic-settings = "^2.9.1"
cisco-outshift-agent-utils = "^0.0.2"
coloredlogs = "^15.0.1"
langgraph-supervisor = "^0.0.20"
langchain-openai = "^0.3.16"
langchain-anthropic = "^0.3.12"
agntcy-acp = "^1.1.3"
httpx-sse = "^0.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
